Made additional edits
