+++ src/components/matchHistory/MatchHistory.tsx
@@ -6,9 +6,7 @@ import { getMatches } from "../../redux/statsSelectors";
 import { useSelector } from "react-redux";
 import { Loading } from "../Loading";
 import { useNavigate } from "react-router-dom";
-import { Player } from "../../types/domain/Player";
-import { Match } from "../../types/domain/Match";
-import { match } from "assert";
+import { matchesToPlayers } from "../../redux/statsReducer";
 
 export const MatchHistory = React.memo(function MatchHistory() {
     const navigate = useNavigate();
@@ -41,32 +39,3 @@ export const MatchHistory = React.memo(function MatchHistory() {
         </Flex>
     );
 });
-
-// ADDITIONAL TEST CODE BELOW (AXIOMATICADI)
-// given a collection of matches, return all of the players in those matches
-function matchesToPlayers(matches: Match[]): { [name: string]: Player } {
-    const playerDictionary: { [name: string]: Player } = {};
-    for (const currentMatch of matches) {
-        // iterate through each player, and add those commanders to our commander dictionary
-        for (const player of currentMatch.players) {
-            const currentPlayerName = player.name;
-            const potentialPlayerObj = playerDictionary[currentPlayerName];
-
-            // if the entry doesn't exist, add to dictionary
-            if  (potentialPlayerObj === undefined) {
-                playerDictionary[currentPlayerName] = {
-                    name: currentPlayerName,
-                    matches: [currentMatch],
-                    wins: (player.rank === "1") ? 1 : 0,
-                }
-            } else {
-                // since this player exists, update the currentMatch count
-                playerDictionary[currentPlayerName].matches.push(currentMatch);
-                if (player.rank === "1") { 
-                    playerDictionary[currentPlayerName].wins++; 
-                }
-            }
-        }
-    }
-    return playerDictionary;
-}
