+++ src/redux/statsReducer.ts
@@ -67,31 +67,30 @@ function matchesToCommanders(matches: Match[]): { [id: string]: Commander } {
     return commanderDictionary;
 }
 
-// // given a collection of matches, return all of the players in those matches
-// function matchesToPlayers(matches: Match[]): { [name: string]: Player } {
-//     const playerDictionary: { [name: string]: Player } = {};
-//     for (const currentMatch of matches) {
-//         // iterate through each player, and add those commanders to our commander dictionary
-//         for (const player of currentMatch.players) {
-//             const currentPlayerName = player.name;
-//             const potentialPlayerObj = playerDictionary[currentPlayerName];
+// given a collection of matches, return all of the players in those matches
+export function matchesToPlayers(matches: Match[]): { [name: string]: Player } {
+    const playerDictionary: { [name: string]: Player } = {};
+    for (const currentMatch of matches) {
+        // iterate through each player, and add those commanders to our commander dictionary
+        for (const player of currentMatch.players) {
+            const currentPlayerName = player.name;
+            const potentialPlayerObj = playerDictionary[currentPlayerName];
 
-//             // if the entry doesn't exist, add to dictionary
-//             if  (potentialPlayerObj === undefined) {
-//                 playerDictionary[currentPlayerName] = {
-//                     name: currentPlayerName,
-//                     matches: [currentMatch],
-//                     wins: (player.rank === "1") ? 1 : 0,
-//                 }
-//             } else {
-//                 // since this player exists, update the currentMatch count
-//                 playerDictionary[currentPlayerName].matches.push(currentMatch);
-//                 if (player.rank === "1") { 
-//                     playerDictionary[currentPlayerName].wins++; 
-//                 }
-//             }
-//         }
-//     }
-//     return playerDictionary;
-// }
-
+            // if the entry doesn't exist, add to dictionary
+            if  (potentialPlayerObj === undefined) {
+                playerDictionary[currentPlayerName] = {
+                    name: currentPlayerName,
+                    matches: [currentMatch],
+                    wins: (player.rank === "1") ? 1 : 0,
+                }
+            } else {
+                // since this player exists, update the currentMatch count
+                playerDictionary[currentPlayerName].matches.push(currentMatch);
+                if (player.rank === "1") { 
+                    playerDictionary[currentPlayerName].wins++; 
+                }
+            }
+        }
+    }
+    return playerDictionary;
+}
